+++ js/services.js
@@ -1,4 +1,3 @@
-
 function requestCreator (requestType, requestBody) {
   // A request generator body with type of request to perform and the body/data to send to the api handler.
   // getGeoLocation method will be added later